Removed deprecated methods, moved remaining methods from common to resource or ov_client and fixed unit tests for those.
