+++ hpOneView/image_streamer/resources/golden_images.py
@@ -30,9 +30,8 @@ from future import standard_library
 standard_library.install_aliases()
 
 
-from hpOneView.resources.resource import ResourceClient
+from hpOneView.resources.resource import ResourceClient, extract_id_from_uri
 from hpOneView.resources.task_monitor import TaskMonitor
-from hpOneView.common import extract_id_from_uri
 from urllib.parse import quote
