+++ hpOneView/image_streamer/resources/os_volumes.py
@@ -30,8 +30,7 @@ from future import standard_library
 standard_library.install_aliases()
 
 
-from hpOneView.common import extract_id_from_uri
-from hpOneView.resources.resource import ResourceClient
+from hpOneView.resources.resource import ResourceClient, extract_id_from_uri
 
 
 class OsVolumes(object):
