+++ hpOneView/image_streamer/resources/plan_scripts.py
@@ -30,8 +30,7 @@ from future import standard_library
 standard_library.install_aliases()
 
 
-from hpOneView.resources.resource import ResourceClient
-from hpOneView.common import extract_id_from_uri
+from hpOneView.resources.resource import ResourceClient, extract_id_from_uri
 
 
 class PlanScripts(object):
